Added 6 entites and data passed successfully

+++ src/main/java/com/server/ZeroZinema/repository/FeedbackRepository.java
@@ -3,6 +3,6 @@ package com.server.ZeroZinema.repository;
 import com.server.ZeroZinema.model.Feedback;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FeedbackRepository extends JpaRepository<Feedback, Long> {
+public interface FeedbackRepository extends JpaRepository<Feedback, Integer> {
 
 }
