+++ src/main/java/com/server/ZeroZinema/repository/MovieRepository.java
@@ -3,6 +3,6 @@ package com.server.ZeroZinema.repository;
 import com.server.ZeroZinema.model.Movie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MovieRepository extends JpaRepository<Movie, Long> {
+public interface MovieRepository extends JpaRepository<Movie, Integer> {
 
 }
